Refactor imports in purchase and sale pages
